refactor: move search-result to entities
