+++ src/entities/search/search-result.tsx
